feat: Implement macOS build speed improvements and warning-free compilation
- Added requirements for a warning-free build on macOS, ensuring zero clang warnings during engine compilation.
- Established strict compilation rules under `-Wall -Wextra` for engine projects, with minimal suppression documented.
- Fixed defect-indicating warnings in the codebase, including polymorphic deletion safety and format string issues.
- Enhanced reflection macros to maintain platform-specific warning scopes without altering type layouts.
- Resolved include hygiene warnings without modifying vendored code, ensuring proper casing and header inclusion.
- Introduced tasks for implementing the above requirements, including reflection macro fixes, defect resolution, and include hygiene.
- Documented macOS build performance baselines and configurations in the knowledge base, detailing improvements in build times and object sizes.
- Updated `premake5.lua` to support new debug info options and ensure macOS-specific configurations do not affect Windows builds.

+++ engine/editor/source/gui.cpp
@@ -96,10 +96,10 @@ void EditorGUI::draw() {
 
 			if (ImGui::BeginMenuBar()) {
 				if (ImGui::BeginMenu("resolution")) {
-					for (int i = 0; i < ARRAY_LENGTH(g_viewport_resolutions); ++i) {
+					for (int i = 0; i < (int)ARRAY_LENGTH(g_viewport_resolutions); ++i) {
 						if (ImGui::MenuItem(
 							(std::to_string(g_viewport_resolutions[i][0]) + "x" + std::to_string(g_viewport_resolutions[i][1])).c_str(),
-							nullptr, nullptr, i != m_current_resolution)) {
+							nullptr, nullptr, i != (int)m_current_resolution)) {
 							m_current_resolution = i;
 						}
 					}
